fix: plugin refactor v1 (#121)
* fix: refactor plugin directory

* fix: refactor plugin redux store
