+++ src/components/Plugins/NodeInfo/NodeInfoDot.js
