+++ src/components/Plugins/PluginView/index.js
