+++ src/components/Plugins/Terminal/TerminalInput.js
