+++ src/components/Plugins/Terminal/index.js
